+++ packages/sdk/src/types/lend/index.ts
@@ -1,3 +1,2 @@
 export * from '@/types/lend/base.js'
-export * from '@/types/lend/contracts.js'
 export * from '@/types/lend/morpho.js'
